fixed divide by 0 bug

+++ Backend/src/main/kotlin/com/cardcounting/backend/dto/StrategyRequest.kt
@@ -5,6 +5,6 @@ data class StrategyRequest(
     val standsOnSoft17: Boolean,
     val bankroll: Int,
     val minBetSize: Int,
-    val minBetIncrement: Int,
+    val minBetIncrement: Double,
     val betSpread: Int
 )
